refactor(proj): move Pep/10 to separate file
